added changeColor
refer to task: "Adapt your script such that the laser points of the file
"Reflector.asc" are assigned a red
color (i.e. ASCII line: X Y Z 255 0 0). For this you read the file
line-by-line and replace the
color information by defining it as red: “255 0 0”, e.g"

--> def changeColor with args changeFile, r, g, b
is called when file is Reflector.asc